% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_gc.R
\name{seq_gc}
\alias{seq_gc}
\title{Get the GC content of a nucleic acid sequence}
\usage{
seq_gc(seq, digits = 4)
}
\arguments{
\item{seq}{A string - the nucleic acid sequence formatted as a single string}

\item{digits}{An integer - rounding precision for the result (see [base::round()])}
}
\value{
A double, the GC content of *seq*, with precision defined by *digits*
}
\description{
Get the GC content of a nucleic acid sequence
}
