% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_revcomp.R
\name{seq_revcomp}
\alias{seq_revcomp}
\title{Get the reverse complement of a nucleic acid sequence}
\usage{
seq_revcomp(seq)
}
\arguments{
\item{seq}{A string - the nucleic acid sequence formatted as a single string}
}
\value{
A string - the reverse complement of *seq*
}
\description{
Get the reverse complement of a nucleic acid sequence
}
